+++ __init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import
+
+#from packages.pymybase.mylogginbase import MyLoggingBase
+
+__all__ = ( , )