+++ CHANGELOG.md
@@ -3,4 +3,7 @@
 - this file - CHANGELOG.md
 - myinputdata.py class to (de) serialize each input line
 ## Changed
-- refactored fallowing pylint
+- refactored following pylint
+- pymybase located in packages module
+## Fixed
+## Removed